Stupid attempt to fix the stupid Goko load timing problem.
Goko's load process is so fragile that the #how-to-play element can get used before it is initialized. This is definitely a race condition, so no promises about reproducing it exactly, but it happens consistently on my machine in Firefox if I give the browser any real work to do at the DOMContentLoaded event.

I added a brief interval (100ms) to prevent this from happening, but this is certainly not a complete solution. I expect that some users will run into the same problem anyway.

I could add a longer interval, but that results in the avatars.js module not loading correctly. That module alters code in FS.Templates.Launch screen that only gets read on initialization. So unless that change happens quickly enough, the avatars module has no effect.